Fix all the issues
