+++ services/absence_service.py
@@ -57,7 +57,7 @@ class AbsenceService:
             int: The minimum semester.
         """
         try:
-            absencesPage = self.getAbsencesPage()
+            absencesPage = self.oge_scraper.getAbsencesPage()
             soup = BeautifulSoup(absencesPage, 'html.parser')
             min_semester_text = soup.find('span', class_='ui-menuitem-text').get_text()
             min_semester = int(min_semester_text.split()[-1])
@@ -77,7 +77,7 @@ class AbsenceService:
             int: The maximum semester.
         """
         try:
-            absencesPage = self.getAbsencesPage()
+            absencesPage = self.oge_scraper.getAbsencesPage()
             soup = BeautifulSoup(absencesPage, 'html.parser')
             max_semester_text = soup.find_all('span', class_='ui-menuitem-text')[-1].get_text()
             max_semester = int(max_semester_text.split()[-1])
@@ -97,29 +97,8 @@ class AbsenceService:
             str: The absences page.
         """
         try:
-            response = self.session.post(self.absences_url, headers=self._get_headers(), data=self._get_absences_data(semester), timeout=5)
-            return self._extract_content(response)
+            return self.oge_scraper.postAbsencesForSemester(semester)
         except Exception as e:
             logging.error(f"Error in selecting absences semester: {e}")
             return None
     
-    
-    def _get_absences_data(self, semester):
-        """
-        This method is used to get the data for the request.
-
-        Parameters:
-            semester (int): The semester to select.
-
-        Returns:
-            dict: The data.
-        """
-        return {
-            "javax.faces.partial.ajax": "true",
-            "javax.faces.source": "ficheEtudiantForm:j_id_16_" + str(semester),
-            "javax.faces.partial.execute": "@all",
-            "javax.faces.partial.render": "ficheEtudiantForm:panel",
-            "ficheEtudiantForm:j_id_16_" + str(semester): "ficheEtudiantForm:j_id_16_" + str(semester),
-            "ficheEtudiantForm_SUBMIT": "1",
-            "javax.faces.ViewState": "0"
-        }
