+++ services/oge_scraper.py
@@ -15,10 +15,10 @@ class OgeScraper:
         Returns:
         None
         """
-        self.home_url = "https://oge.polytechnique.fr/ogesip/"
-        self.login_url = "https://oge.polytechnique.fr/ogesip/login.jsp"
-        self.absences_url = "https://oge.polytechnique.fr/ogesip/absences.jsp"
-        self.grades_url = "https://oge.polytechnique.fr/ogesip/notes.jsp"
+        self.home_url = "https://iutdijon.u-bourgogne.fr/oge-esirem/"
+        self.login_url = "https://casiut21.u-bourgogne.fr/cas-esirem/login?service=https%3A%2F%2Fiutdijon.u-bourgogne.fr%2Foge-esirem%2F"
+        self.absences_url = "https://iutdijon.u-bourgogne.fr/oge-esirem/stylesheets/etu/absencesEtu.xhtml"
+        self.grades_url = "https://iutdijon.u-bourgogne.fr/oge-esirem/stylesheets/etu/bilanEtu.xhtml"
             
 
     def get(self, url):
@@ -32,7 +32,7 @@ class OgeScraper:
             requests.Response: The response of the request.
         """
         print(f"Get {url}")
-        return self.session.get(url)
+        return self.session_manager.session.get(url)
 
     def getAbsencesPage(self):
         """
@@ -45,7 +45,7 @@ class OgeScraper:
             str: The absences page.
         """
         print("Get absences page...")
-        response = self.session.get(self.absences_url)
+        response = self.session_manager.session.get(self.absences_url)
         return response.text
     
     def getGradesPage(self):
@@ -59,7 +59,7 @@ class OgeScraper:
             str: The grades page.
         """
         print("Get grades page...")
-        response = self.session.get(self.grades_url)
+        response = self.session_manager.session.get(self.grades_url)
         return response.text
     
     def _get_headers(self):
@@ -77,6 +77,21 @@ class OgeScraper:
             "X-Requested-With": "XMLHttpRequest",
             "Content-Type": "application/x-www-form-urlencoded; charset=UTF-8"
         }
+    
+    def postAbsencesForSemester(self, semester):
+        """
+        This method is used to select the absences semester.
+
+        Parameters:
+            semester (int): The semester to select.
+
+        Returns:
+            str: The absences page.
+        """
+        print(f"Select semester {semester}...")
+        data = self._get_absences_data(semester)
+        response = self.session_manager.session.post(self.absences_url, data=data, headers=self._get_headers())
+        return self._extract_content(response)
 
     def _extract_content(self, response):
         """
@@ -90,3 +105,23 @@ class OgeScraper:
         """
         content = response.text.split("![CDATA[")[1].split("]]")[0]
         return content
+    
+    def _get_absences_data(self, semester):
+        """
+        This method is used to get the data for the request.
+
+        Parameters:
+            semester (int): The semester to select.
+
+        Returns:
+            dict: The data.
+        """
+        return {
+            "javax.faces.partial.ajax": "true",
+            "javax.faces.source": "ficheEtudiantForm:j_id_16_" + str(semester),
+            "javax.faces.partial.execute": "@all",
+            "javax.faces.partial.render": "ficheEtudiantForm:panel",
+            "ficheEtudiantForm:j_id_16_" + str(semester): "ficheEtudiantForm:j_id_16_" + str(semester),
+            "ficheEtudiantForm_SUBMIT": "1",
+            "javax.faces.ViewState": "0"
+        }
